rag-chatbot/pyproject.toml
[project]
name = "rag-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cohere>=5.20.0",
    "python-dotenv>=1.2.1",
    "qdrant-client~=1.7.0",
    "requests>=2.32.5",
    "trafilatura>=2.0.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.28.0",
    "streamlit>=1.32.0",
]
